+++ ChangeLog.md
@@ -1,3 +1,8 @@
+###### v1.3.0 (2020-12-10):
+ * Added handling of sequence start value to JS.
+ * Added verification of sequence start input.
+ * Added verification of sequence start and length input values.
+
 ###### v1.2.7 (2020-12-10):
  * Added sequence start to Input Summary Table.